==> hemnet-scrape/sthlm/3. Sthlm -  Clean_get sthlm_raw_clean.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import re
import math

# Intro

In this notebook I´m data-cleaning the csv created in "2. Sthlm..." 

data = pd.read_csv("sthlm_raw.csv")

data.info()

# Drop rows where "kvm" or "rum" is missing

Start with dropping for efficiency (avoiding the risk of spending time fixing rows I'll later drop)

kvm and rum are two of the most important values and I prefer dropping rows where they are missing from any other approach. 

I feel ok with dropping since I'm only loosing app. 200 / 58000 rows ~ 0.35% of the total sample

data = data.loc[~data.rum.isna()]

data = data.loc[~data.kvm.isna()]

data.info()

# Adapting numerical decimals to pandas

# Changing , to . for kvm
data.kvm = data.kvm.str.replace(",", ".")

# Changing , to . for rum
data.rum = data.rum.str.replace(",",".")

# Data-types

## kvm

# kvm changed to float from object
data.kvm = data.kvm.astype(float)

## rum

# rum changed to float from object
data.rum = data.rum.astype("float")

## datum

# Need to remove and replace text before converting to datetime-type

# remove text "Såld" from all rows
data.datum = data.datum.str.replace("Såld", "")

# Function to change text for month to numerical value for month 
def manader_to_num(datum): 
    if "januari" in datum: 
        return datum.replace("januari", "1")
    elif "februari" in datum: 
        return datum.replace("februari", "2")
    elif "mars" in datum: 
        return datum.replace("mars", "3")
    elif "april" in datum: 
        return datum.replace("april", "4")
    elif "maj" in datum: 
        return datum.replace("maj", "5")
    elif "juni" in datum: 
        return datum.replace("juni", "6")
    elif "juli" in datum: 
        return datum.replace("juli", "7")
    elif "augusti" in datum:
        return datum.replace("augusti", "8")
    elif "september" in datum: 
        return datum.replace("september", "9")
    elif "oktober" in datum: 
        return datum.replace("oktober", "10")
    elif "november" in datum: 
        return datum.replace("november", "11")
    elif "december" in datum: 
        return datum.replace("december", "12")
    else: 
        return np.nan

# apply function to change month from text to numerical
data.datum = data.datum.apply(manader_to_num)

# change type from object to datetime. I'm specifying that it should parse day first 
data.datum = pd.to_datetime(data.datum, dayfirst=True)

data.info()

# NaN in avgift: Future update (maybe?)

## Description for future update

I'm not really interested in the values in this column so i don't think I'll bother filling the missing values. 

When there are values missing it's simply because that info has been left out from my data-source (Hemnet)

# NaNs in gata_lst: Future update
## Description for future update: gata_lst - fillna from street name in adress column

# Inspecting rows there gata_lst is NaN
# I did not pick up all special cases when I created the data source for gata_lst
# i.e.: é 
data.loc[data.gata_lst.isna()]

# Outliers

## kvm

data.kvm.describe()

### min

data.kvm.nsmallest()

# Values look reasonable 
data.loc[7394]

https://www.hemnet.se/salda/lagenhet-1rum-kungsholmen-fridhemsplan-stockholms-kommun-hantverkargatan-77,-5-tr-696839

Found the object and the information seems to be correct

### max

data.kvm.nlargest()

# value for kvm is not in sync with other values
# almost sure that it should be 40.5 kvm
data.loc[37355]

# changing the value for index = 37355
data.loc[37355, "kvm"] = 40.5

# value for kvm is in sync with other values
data.loc[29779]

# value for kvm in sync with other values
data.loc[29388]

# moving on...

## rum

# Looks reasonable
data.rum.describe()

# Looks reasonable
data.rum.value_counts()

## Slutpris

# changing format for displaying floats
pd.options.display.float_format = '{:f}'.format

# inspecting column slutpris
# max seems reasonable
# min is wrong
data.slutpris.describe()

### min

data.slutpris.nsmallest()

# looking at nsmallest(1)
# can't infer appropriate value from other values
# removing this from data-set
data.loc[56366]

# removing row
data.drop(56366, inplace = True)

# looking at nsmallest(2)
# Probably missed a 0 for slutpris
# More efficient to just remove this row from the data-set
data.loc[40973]

# remove row
data.drop(40973, inplace = True)

# looking at nsmallest(3)
# values seem reasonable
data.loc[51623]

### max

# values seem reasonable
data.slutpris.nlargest()

## Datum

# Looks good
data.datum.describe()

## Prisförändring

# Inspecting prisförändring

# min-value is suspicious
# max-value is suspicious
data.prisförändring.describe()

### min

data.prisförändring.nsmallest()

# looks strange but not impossible
data.loc[54559]

https://www.hemnet.se/salda/lagenhet-3rum-sodermalm-mariatorget-stockholms-kommun-rosenlundsgatan-28b-531405 

Info seems to be correct

### max

# nlargest(1) and (2) is for sure wrong
data.prisförändring.nlargest()

# removing top2 
data.drop(50144, inplace = True)
data.drop(32232, inplace = True)

# unreasonable
data.loc[45062]

data.drop(45062, inplace = True)

# possible but not likely to be correct
data.loc[14701]

data.drop(14701, inplace = True)

data.prisförändring.nlargest()

# possible 
data.loc[7192]

## avgift

# min-value: some apartments have zero fee
# max-value is not unreasonable
data.avgift.describe()

data.avgift.nlargest()

# reasonable
data.loc[30351]

# Final inspection

# I'm satisfied with the data-types
data.info()

# No values seem unreasonable now
data.describe()

# Resetting index before writing to csv

# resetting index and dropping the outdated index
data.reset_index(inplace = True, drop=True)

# inspecting index, looks correct
data

# Clean data to_csv

data.to_csv("sthlm_raw_clean.csv", index = False)
