==> hemnet-scrape/sthlm/4. Sthlm - Add_get sthlm_raw_clean_added_columns.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import re
import math

data = pd.read_csv("sthlm_raw_clean.csv")

data

# Intro

In this file I'm adding columns from the existing data. 

To clarify I'm not adding new data but adding new columns based on existing data.

Columns to add:

1) kr_kvm

2) år

3) månad

4) år-månad

# Add kr_kvm

# adding kr_kvm based on price and square-metres
data["kr_kvm"] = data.slutpris / data.kvm

# change type from float to int
data.kr_kvm = data.kr_kvm.astype(int)

## Inspection values kr_kvm

Taking the opportunity to check if there are unreasonable values based off price / square-metres

data.kr_kvm.describe()

# Looks reasonable
data.kr_kvm.nsmallest()

# Looks reasonable
data.kr_kvm.nlargest()

# Add år

# adding year-column for future slicing
data["år"] = pd.DatetimeIndex(data["datum"]).year

# Add månad

data["månad"] = pd.DatetimeIndex(data["datum"]).month

# Add år-månad

data["år_månad"] = pd.to_datetime(data.år.astype(str) + "-" + data.månad.astype(str))

# Inspection

# looks good 
data.head()

# looks good
# notice though that i'm failing to keep the datetime-format for the column "datum" when reading the csv
data.info()

# write to csv

I've still got one more round of cleaning before moving on to analysis. 

I'm therefore still naming my csv-file in a way so i can keep track of the changes I've made

data.to_csv("sthlm_raw_clean_added_columns.csv")
